requirements.txt
khl.py
tortoise-orm[asyncmy]
PyYAML